Simplify placing item in the sidebar by moving the output area node directly and using CSS to hide unwanted sub parts
